+++ src/Strongweak/Strength.hs
@@ -5,7 +5,7 @@ module Strongweak.Strength where
 import Strongweak.Weaken ( type Weakened )
 import Data.Kind ( type Type )
 
-import Strongweak.WeakenN
+import Strongweak.Chain
 import GHC.TypeNats ( type Natural )
 
 -- | Strength enumeration: it's either strong, or weak.
@@ -30,5 +30,5 @@ type family SW (s :: Strength) a :: Type where
     SW Strong a =          a
     SW Weak   a = Weakened a
 
--- | Shortcut for a 'SW'-wrapped 'WeakenN'.
-type SWN (s :: Strength) (n :: Natural) a = SW s (WeakenN n a)
+-- | Shortcut for a 'SW'-wrapped 'SWChain'.
+type SWN (s :: Strength) (n :: Natural) a = SW s (SWChain n a)